[get-scores] 점수 받아오기 성공

+++ data/src/main/java/com/capstone2/data/model/session/GetScoresResponseDTO.kt
@@ -7,7 +7,7 @@ data class GetScoresResponseDTO(
     @SerializedName("date")
     val date: String,
     @SerializedName("finalScore")
-    val finalScore: Int,
+    val finalScore: Double,
     @SerializedName("sessionId")
     val sessionId: Int
 )
