+++ domain/src/main/java/com/capstone2/domain/model/session/GetScoresResult.kt
@@ -2,6 +2,6 @@ package com.capstone2.domain.model.session
 
 data class GetScoresResult(
     val date: String,
-    val finalScore: Int,
+    val finalScore: Double,
     val sessionId: Int
 )
